showcase/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-socketio[asyncio_server]>=5.11.0
pylsl>=1.16.2
python-multipart>=0.0.9
